requirements.txt
edge-tts>=6.1.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
openai-whisper>=20231117
moviepy>=2.0.0
Pillow>=10.0.0
